backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0

# HTTP and async support
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0

# Data processing and ML
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.8.0
scikit-learn>=1.3.0

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment==3.3.2
transformers>=4.36.0
torch>=2.1.0
langdetect>=1.0.9

# Text processing utilities
geotext>=0.4.0

# API integrations
tweepy>=4.14.0
praw>=7.7.1
telethon>=1.40.0
feedparser>=6.0.0

# Rate limiting and monitoring
slowapi>=0.1.9
schedule>=1.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0

# Additional utilities
python-multipart>=0.0.6 